remove tracing_actor, pull tracing subscriber into its own module
